requirements.txt
coverage==4.5.1
Django==2.1.2
django-appconf==1.0.2
django-compressor==2.2
django-debug-toolbar==1.10.1
django-elasticsearch-dsl==0.5.0
django-extensions==2.1.3
django-filter==2.0.0
django-js-asset==1.1.0
django-mptt==0.9.1
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
elasticsearch==6.3.1
elasticsearch-dsl==6.1.0
ipaddress==1.0.22
mod-wsgi==4.6.4
mysqlclient==1.3.13
Pillow==5.3.0
psycopg2-binary==2.7.5
Pygments==2.2.0
PyJWT==1.6.4
python-dateutil==2.7.3
python-dotenv==0.9.1
pytz==2018.5
rcssmin==1.0.6
redgreenunittest==0.1.1
rjsmin==1.0.12
selenium==3.14.1
six==1.11.0
sqlparse==0.2.4
urllib3==1.23
